extract page controls to new component
