+++ src/App.tsx
@@ -1,27 +1,17 @@
-import {
-  useCallback,
-  useDeferredValue,
-  useEffect,
-  useId,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { AboutMe } from "./AboutMe";
 import "./App.css";
 import { Education } from "./Education";
 import { Experience } from "./Experience";
 import { Links } from "./Links";
-import { Select } from "./Select";
-import { SelectMobile } from "./SelectMobile";
+import { PageControls } from "./PageControls";
 import { ThemeContext } from "./ThemeContext";
 import { useLocalStorage } from "./useLocalStorage";
 import { useSystemPreferredTheme } from "./useSystemPreferredTheme";
-import { useWindowWidth } from "./useWindowWidth";
 
 function App() {
   const htmlRef = useRef(document.documentElement);
-  const themePickerLabel = useId();
+
   const [activeTheme, setActiveTheme] = useState<"dark" | "light">("dark");
 
   // use local storage persisting state for app-level theme preference.
@@ -31,13 +21,7 @@ function App() {
     "auto"
   );
 
-  const windowWidth = useDeferredValue(useWindowWidth());
   const systemPreferredTheme = useSystemPreferredTheme();
-
-  const isAndroid = useMemo(
-    () => navigator.userAgent.toLowerCase().includes("android"),
-    []
-  );
 
   // set the site theme using a class on the main html element
   const setTheme = useCallback((dark: boolean) => {
@@ -67,46 +51,10 @@ function App() {
 
   return (
     <ThemeContext.Provider value={activeTheme}>
-      <div
-        className={`flex items-center gap-4 fixed m-4 p-2 rounded 
-          print:invisible sm:shadow-lg sm:bg-background sm:top-0 right-0 max-sm:bottom-0`}
-      >
-        <div className="flex items-baseline gap-2 relative">
-          <label
-            className={`max-sm:text-sm max-sm:absolute py-1 px-2 max-sm:left-2 rounded
-              max-sm:top-[-0.75rem] bg-background`}
-            id={themePickerLabel}
-          >
-            Theme
-          </label>
-          {windowWidth < 640 ? (
-            <SelectMobile
-              className="shadow-lg"
-              aria-labelledby={themePickerLabel}
-              options={["auto", "dark", "light"]}
-              selectedOption={themePreference ?? "auto"}
-              setSelectedOption={setThemePreference}
-            />
-          ) : (
-            <Select
-              position={"bottom"}
-              aria-labelledby={themePickerLabel}
-              options={["auto", "dark", "light"]}
-              selectedOption={themePreference ?? "auto"}
-              setSelectedOption={setThemePreference}
-            />
-          )}
-        </div>
-        {!isAndroid && (
-          <button
-            className={`material-symbols-outlined text-4xl bg-background rounded-full max-sm:p-2 
-              border border-neutral-400 dark:border-white max-sm:shadow-lg`}
-            onClick={() => window.print()}
-          >
-            print
-          </button>
-        )}
-      </div>
+      <PageControls
+        themePreference={themePreference}
+        setThemePreference={setThemePreference}
+      />
       <div className="space-y-4 max-sm:mb-16">
         <div className="flex justify-center">
           <main className="max-w-4xl px-4 flex-grow space-y-4 mb-8">
